PROJFBPRO-69: Rename get_the_state_of_the_system.py to ea_system_prerequisites_functions.py
